move adapters out of gui class
